Move message types to separate module
