+++ magic-wormhole.cabal
@@ -35,6 +35,7 @@ library
     , unordered-containers
     , websockets
   exposed-modules:
+      MagicWormhole.Internal.Messages
       MagicWormhole.Internal.Rendezvous
       MagicWormhole.Internal.WebSockets
   default-language: Haskell2010
@@ -69,6 +70,6 @@ test-suite tasty
     , tasty-hedgehog
     , tasty-hspec
   other-modules:
-      Rendezvous
+      Messages
       WebSockets
   default-language: Haskell2010
